pyproject.toml
[tool.poetry]
name = "pancaketrade"
version = "0.7.1"
description = "Trading bot for PancakeSwap"
authors = ["Valentin Bersier <vbersier@gmail.com>"]

[tool.poetry.scripts]
trade = 'pancaketrade:main'

[tool.poetry.dependencies]
python = ">3.7.1,<3.10"
requests = "^2.25.1"
importlib-metadata = {version="^4.0.1", python="<3.8"}
loguru = "0.*"
python-telegram-bot = "^13.10"
web3 = "^5.18.0"
cachetools = "^4.2.2"
PyYAML = "^5.4.1"
click = "^7.1.2"
yamale = "^3.0.6"
questionary = "^1.9.0"
peewee = "^3.14.4"
apscheduler = "<4.0.0"
python-dotenv = "0.*"

[tool.poetry.dev-dependencies]
pre-commit = "^2.17.0"
flake8 = "^4.0.1"
flake8-bugbear = "^22.1.11"
flake8-comprehensions = "^3.8.0"
black = { version = "*", allow-prereleases = true }
mypy = ">=0.931"
types-PyYAML = "^6.0.4"
types-requests = "^2.27.9"
types-cachetools = "^4.2.9"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
